+++ rogues/rogue.py
@@ -1,4 +1,4 @@
-from pki import get_kp, pubkey_to_addr, sign, verify
+from utils.pki import get_kp, sign
 import requests
 import json
