+++ utils/helpers.py
@@ -1,0 +1,42 @@
+import pytz
+import json
+import socket
+import random
+import hashlib
+from datetime import datetime
+from utils.validation import validate_difficulty
+
+
+def utcnow():
+    return int(datetime.now(tz=pytz.utc).timestamp())
+
+
+def remap(mapping):
+    return [{'key': k, 'value': v} for k, v in mapping.items()]
+
+
+def resolve(ip):
+    return socket.gethostbyaddr(ip)[0]
+
+
+# Encode dicts (messages loaded from JSON for example) in standard way
+def standard_encode(dictionary):
+    return bytes(
+        json.dumps(dictionary, sort_keys=True, separators=(',', ':')),
+        'utf-8')
+
+
+def hasher(dictionary):
+    return hashlib.sha256(standard_encode(dictionary)).hexdigest()
+
+
+# TODO: Do this in C or other efficient lib..
+def mine(content=None):
+    nonce = random.randrange(config['max_randint'])
+    while True:
+        content['nonce'] = nonce
+        hashed = hasher(content)
+        if validate_difficulty(hashed):
+            break
+        nonce += random.randrange(config['nonce_jump'])
+    return hashed, nonce
